tools/conversate/client/android/requirements.txt
kivy
python_dotenv
Cython==0.29.33
virtualenv
pyjnius